[Feature][Philippe] Split lexer-parser, added symbol table, fix bugs
