+++ src/compiler/test.kyut
@@ -1,0 +1,18 @@
+nom   x    <3nom y <3
+stwing z  <3
+stwing a <3
+
+
+pwease y <3
+x + 5 <3
+x + 5 +4+2+x+y <3
+x + 3*5 +4<3 y +     3*5 +4<3
+
+y= 3*2 +5 <3 
+x = 2+3*5-3+7 <3 
+z = "hatdog" <3 	
+a = "ham<3borgs" <3
+a = "hamsdfaksdjhf  lkjsdhaflkj lkdj  jasdljfh" <3
+z = a <3
+
+rawr 5 * (4 + 5) * (4 + (6*7) - 5)<3
